+++ test/yesoul_s3_test.dart
@@ -159,7 +159,7 @@ void main() {
         ),
       ),
     ]) {
-      final sum = testPair.data.fold<double>(0.0, (a, b) => a + b);
+      final sum = testPair.data.fold<int>(0, (a, b) => a + b);
       test("$sum ${testPair.data.length}", () async {
         final bike = DeviceFactory.getYesoulS3();
         bike.initFlag();
